backend/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
openai==1.3.7
python-dotenv==1.0.0
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0